testSearcherCorrectReferencesAfterRestore
